requirements.txt
# Otomoto Parser – Dependencies

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
playwright>=1.44.0

# Storage
sqlalchemy>=2.0.0

# Settings & config
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Export
openpyxl>=3.1.0

# CLI & logging
typer[all]>=0.9.0
rich>=13.0.0

# Retry logic
tenacity>=8.2.0
